Support dots in repo names (#21)
Having a dot is supported in GitHub, while the `SHORT` regex is too
restrictive and disallows this. Support is added for repositories that
contain a period or full stop.

There is no definitive guide to allowed characters for GitHub
repositories, but you can piece together the allowed characters via
moby/moby#679 and desktop/desktop#3090.

Tests were also added that verified the allowed characters in the
repository name.