Fix uploading reviews on custom GH URLs
octocrab does a `Url::join` on the base_url and the custom endpoint I'm
specifying. `Url::join` is a bit tricky with its semantics. Turns out
providing a leading `/` causes the existing URL path to be wiped out.

Provide a relative path so it's simply appended like we expect. See [0]
for example.

This closes #31.

[0]: https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=3665b8c75f44cd0d4ade63e416084f8b