Allow statically linking openssl
This is mostly to allow hassle-free installation on NixOS and such.
Attentive reader might ask: why not rustls then? At the time of writing,
octocrab and reqwest support rustls, but git2 does not. It might be a
good idea to try to switch to gitoxide in lie of git2, but that's going
to be a larger change!