Download PR description on "get" as well
Having the PR description is useful when doing a review, if only to see
whether the code solves the problem that the PR is (hopefully) addressing.

We download the PR description (if available) and save it in the review
file. The PR description is ignored when checking the status or the
validity of a review. In the current implementation we rely on there
not being any quoted text in the review part of the review file. We also
require the diff header to be present in order to be able to distinguish
between the review header and the actual diff content.

Unfortunately this means that this diff header line cannot be removed
when using snips.