Download PR description on "get" as well
Having the PR description is useful when doing a review, if only to see
whether the code solves the problem that the PR is (hopefully) addressing.

We download the PR description (if available) and save it in the review
file as quoted text. The PR description both composes with snips (as
snips are resolved before parsing) and is checked for corruption.