Implement snips
Snips are a way to cut out noise from PR review. For example, if the PR
is really long but 90% of it looks good, you can snip a contiguous
section of quoted text by deleting the lines and inserting a:

    line 1
    [...]
    line 5

The `[...]` tells prr to replace the snipped lines with the original
text. No information is lost if a snip is used.

This should be very familiar to folks who have used mailing lists
before.