src/prr: ignore untracked files explicitly for "apply"
The git2 documentation states that untracked files shouldn't be included
by default when calling "statuses" without a "StatusOption".

Unfortunately this doesn't seem to be the case in my testing. Untracked
files are causing the statuses not to be empty which results in the
command bailing out.

We fix this by setting the include_ignore option to false.