src/prr: add test for apply_pr
We create a repo in a temp directory, add a non-tracked file and then
check whether we can run apply_pr with a saved review on it.

Since we want to be able to appy PR diffs even if there are untracked
files in the repo, this should succeed at all times.