Adds support for providing a GitHub token via envvar (#75)
Adds support for providing a GitHub token via an environment variable:

- **Envvar passed in config**: if there's a string value passed to `token` that begins with `$`, treat it as an envvar and perform a lookup
- **Known envvar present in environment**: As per envvars in https://cli.github.com/manual/gh_help_environment, if there's no value set for `token` then look for `GH_TOKEN`, `GITHUB_TOKEN`, `GH_ENTERPRISE_TOKEN`, or `GITHUB_ENTERPRISE_TOKEN`, in that order, and use the first token found.
- **Any other string**: treat it like a token.

I've also added a whole bunch of tests for as many scenarios/permutations of this that I can think of, as well as added `serial_test` as a dev dependency so the envvar tests run sequentially to avoid envvar definitions/removals clobbering each other. Please let me know if I've missed anything!

Closes #61